+++ lib/cleon/gateway.rb
@@ -1,0 +1,9 @@
+require_relative "arguards"
+
+module Cleon
+
+  class Gateway
+    include Cleon::ArGuards
+  end
+
+end